Add initial CMake build system
This is not yet supported for production builds but can already produce working
binaries on OS X and Linux with clang and gcc.

The intention is to improve support to the point where it can integrate with
the LLVM runtime platform, cover all platforms, runtime/release build
configurations and run the tests.

Patch by Jack Howarth!

git-svn-id: https://llvm.org/svn/llvm-project/openmp/trunk@209994 91177308-0d34-0410-b5e6-96231b3b80d8