Re-indent the CMake refactor to two-space indention
I apologize for this nasty commit, but I somehow overlooked Chandler's
comment to re-indent these files to two space indention.  I know this
is a horrible commit, but I figured if it was done quickly after the 
first one, not too many conflicts would arise.

Again, I'm sorry and won't do this again.


git-svn-id: https://llvm.org/svn/llvm-project/openmp/trunk@242301 91177308-0d34-0410-b5e6-96231b3b80d8