Fix alignment in teams-reduction.c test
The runtime will use the global kmp_critical_name as a lock and
tries to atomically store a pointer in there. This will fail
if the global is only aligned by 4 bytes, the size of one int32_t
element. Use a union to ensure the global is aligned to the size
of a pointer on the current platform.

git-svn-id: https://llvm.org/svn/llvm-project/openmp/trunk@319811 91177308-0d34-0410-b5e6-96231b3b80d8